Extract the shared bench generator
breadth-equivalence.ts and batch-equivalence.ts were 375 of their
406 and 420 lines identical — the xorshift rng, the config and
tuple builders, buildStore, the coverage counters, the known-error
set, the seed loop and the whole summary tail.

Two copies of a generator that must agree, with nothing enforcing
agreement, is the same hazard as a harness that reports clean over
a path it never enters. Both files have twice been edited
identically since they arrived, each time checked afterwards with
diff run by hand; that is a habit, not a mechanism, and the first
person to edit only one of them gets a harness that has quietly
stopped testing what its sibling tests. They were deliberately
left separate when they arrived, so that a structural refactor
would not ride along with the behaviour change beside it. That
reason has expired.

harness.ts now owns the generator and exports only what a driver
actually reaches for: the scenario stream, withTimeout, outcome,
the known-error set and the summary tail. The scenario stream
yields stores already instrumented, which is the part that earns
its keep — a harness that builds a store and forgets to wrap it
reports clean with admittedUserset 0, and that is not
hypothetical. The five tallies stay local to each driver and are
passed to the tail, so no module owns a mutable bag that any
future harness could write to unnoticed.

Each harness keeps its own phase doc comment, which documents a
different invariant in each, and its own driver loop. The per-case
classifier stays duplicated: breadth compares strings from
outcome, batch formats its own from a checkMany result, and a
shared version would have to take a pre-formatted string. That is
a smaller hazard than the generator and a worse fit for one
commit; it is left in place knowingly, not overlooked.

Behaviour is unchanged and was held to exact numbers rather than
to a passing verdict. Both harnesses are seeded, so their output
is reproducible: before and after, breadth reports 16900 cases and
0 divergences in phase 1, 21 expected divergences and 0 unknown
error classes in phase 2; batch reports 47760, 0, 22 and 0. The
generator-only counts — 2570 userset rows, 3572 direct rows, 1518
conditioned rows — are identical in both harnesses and unchanged
by the extraction, which is the count that would move first if the
generator had drifted.

The bench README now names harness.ts, since its stated reason for
these files existing is that the generator is reviewable in the
repo, and the generator is no longer in either file it named. It
also drops the claim that a run takes a few minutes; breadth takes
about three seconds and batch about six.

No CI job is added. These are randomized differential harnesses
whose red runs need a person to read them, and a per-PR gate that
prints "21 expected divergences" invites a rubber stamp. They are
run by hand before a change to the resolution machinery.
